+++ backend/functions/todo.js
@@ -1,9 +1,7 @@
-const getTodoDB = require("../db/TodoList/getTodoList.js");
-const getTodoCount = require("../db/TodoList/getTodoCount.js");
-const insertTodo = require("../db/TodoList/insertTodo.js");
-const deleteTodo = require("../db/TodoList/deleteTodo.js");
-const updateTodo = require("../db/TodoList/updateTodo.js");
-const distinctColumns = require("../db/TodoList/distinctColumns.js");
+const getTodoDB = require("../db/Todo/getTodoList.js");
+const insertTodo = require("../db/Todo/insertTodo.js");
+const deleteTodo = require("../db/Todo/deleteTodo.js");
+const updateTodo = require("../db/Todo/updateTodo.js");
 
 const statusCode = require("../utils/statusCode.js");
 const errorMessage = require("../utils/errorMessage.js");
@@ -37,11 +35,13 @@ async function postTodoCallback(req, res) {
   try {
     const todo = req.body;
     const result = await insertTodo(todo);
+    console.log("result: ", result);
     const response = {
       id: result[0].insertId,
     };
     return res.status(statusCode.OK).json(response);
   } catch (e) {
+    console.log(e);
     return res.status(statusCode.DB_ERROR).send(errorMessage.DB_ERROR);
   }
 }
@@ -56,15 +56,15 @@ async function deleteTodoCallback(req, res) {
   }
 }
 
-async function getTodoCountCallback(req, res) {
-  try {
-    const result = await getTodoCount();
-    const count = result[0].length;
-    return res.status(statusCode.OK).json(count);
-  } catch (e) {
-    return res.status(statusCode.DB_ERROR).send(errorMessage.DB_ERROR);
-  }
-}
+// async function getTodoCountCallback(req, res) {
+//   try {
+//     const result = await getTodoCount();
+//     const count = result[0].length;
+//     return res.status(statusCode.OK).json(count);
+//   } catch (e) {
+//     return res.status(statusCode.DB_ERROR).send(errorMessage.DB_ERROR);
+//   }
+// }
 
 async function patchTodoCallback(req, res) {
   const { title, content, author } = req.body;
@@ -78,24 +78,10 @@ async function patchTodoCallback(req, res) {
   }
 }
 
-async function distinctColumnsCallback(req, res) {
-  try {
-    const result = await distinctColumns();
-    const response = {
-      data: result[0],
-    };
-    return res.status(statusCode.OK).json(response);
-  } catch (e) {
-    return res.status(statusCode.DB_ERROR).send(errorMessage.DB_ERROR);
-  }
-}
-
 module.exports = {
   getTodoCallback,
-  getTodoCountCallback,
   postTodoCallback,
   validateTodo,
   deleteTodoCallback,
   patchTodoCallback,
-  distinctColumnsCallback,
 };
